+++ src/components/employees/Employees.css
@@ -1,0 +1,15 @@
+.employees {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+
+}
+
+.employee {
+    background-color: hotpink;
+    border-color: 2px dotted fuchsia;
+    color: black;
+    padding: 0.5rem;
+    margin: 1em;
+}
