+++ src/components/nav/CustomerNav.js
@@ -1,0 +1,25 @@
+import { Link, useNavigate } from "react-router-dom"
+import "./NavBar.css"
+
+export const CustomerNav = () => {
+    const navigate = useNavigate()
+
+    return (
+        <ul className="navbar">
+            <li className="navbar__item active">
+                <Link className="navbar__link" to="/tickets">Tickets</Link>
+            </li>
+            {
+                localStorage.getItem("honey_user")
+                    ? <li className="navbar__item navbar__logout">
+                        <Link className="navbar__link" to="" onClick={() => {
+                            localStorage.removeItem("honey_user")
+                            navigate("/", { replace: true })
+                        }}>Logout</Link>
+                    </li>
+                    : ""
+            }
+        </ul>
+    )
+}
+
