+++ src/components/nav/NavBar.js
@@ -1,25 +1,18 @@
-import { Link, useNavigate } from "react-router-dom"
+import { CustomerNav } from "./CustomerNav"
+import { EmployeeNav } from "./EmployeeNav"
 import "./NavBar.css"
 
 export const NavBar = () => {
-    const navigate = useNavigate()
+    const localHoneyUser = localStorage.getItem("honey_user")
+    const honeyUserObject = JSON.parse(localHoneyUser)
 
-    return (
-        <ul className="navbar">
-            <li className="navbar__item active">
-                <Link className="navbar__link" to="/tickets">Tickets</Link>
-            </li>
-            {
-                localStorage.getItem("honey_user")
-                    ? <li className="navbar__item navbar__logout">
-                        <Link className="navbar__link" to="" onClick={() => {
-                            localStorage.removeItem("honey_user")
-                            navigate("/", { replace: true })
-                        }}>Logout</Link>
-                    </li>
-                    : ""
-            }
-        </ul>
-    )
+    if (honeyUserObject.staff) {
+        //return employee views
+        return <EmployeeNav />
+    } else {
+        //return customer views
+        return <CustomerNav />
+
+    }
 }
